Update cloud.py
Added event and  context args to main function. This should allow you to accept the arguments passed by the GCP Cloud Function runtime when your main function is called.

My implementation will print the date and tiem stamp along with the event_id from the GCP environment that was passed to the function.